Correcio tipo dni

+++ TRY_Boletin17-4/src/try_boletin17/pkg4/Metodos.java
@@ -6,7 +6,7 @@ import javax.swing.JOptionPane;
 public class Metodos {
 
      
-    public void encontrar(double dni,Character letra[],int numero[]){
+    public void encontrar(int dni,Character letra[],int numero[]){
         for(int i=0;i<numero.length;i++){
             if(dni==numero[i]){
                 JOptionPane.showMessageDialog(null,"La letra es: "+letra[i]);
